third_party/WebKit/LayoutTests/fast/hidpi/image-srcset-invalid-descriptor.html

<html>
<head>
<script>
    window.targetScaleFactor = 1;
</script>
<script src="resources/srcset-helper.js"></script>
<script src="../js/resources/js-test-pre.js"></script>
<script>
    addEventListener("load", function() {
        shouldBeTrue('document.getElementById("foo").clientWidth==200');
    }, false);
</script>
</head>

<body id="body">
    <div>This test passes if the image below says 1x with a reddish background when the deviceScaleFactor is 1, and if says 2x with a greenish background when the deviceScaleFactor is 2.</div>
    <img id="foo" src="" srcset="resources/blue-100-px-square.png 43q 1x dfs 3dd, resources/green-200-px-square.png 2x">
</body>
</html>
